--- src/main/java/faang/school/projectservice/validate/ErrorResponse.java
@@ -1,26 +1,0 @@
-package faang.school.projectservice.validate;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-
-import java.time.LocalDateTime;
-
-@Data
-@Builder
-@AllArgsConstructor
-public class ErrorResponse {
-
-    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-    private LocalDateTime timestamp;
-
-    private String url;
-    private int status;
-    private String error;
-    private String message;
-
-    public ErrorResponse(String message) {
-        this.message = message;
-    }
-}
